feat: refine weekly rewards and regular drop voting
Phase 11 (mobile): ultra-rare Spin the Wheel prizes (<5% share of total
draw weight - strictly below; exactly 5% keeps normal rendering, rule
centralised in rarity.ts) render as metallic gold with a static
specular gloss and a completely content-free preview: no star, icon,
title, dot or placeholder. The post-spin result still reveals the
actual server-drawn prize. Regular Drop voting moves from the Home
feed to Weekly Drops: a header action (Vote icon, unvoted-badge,
hidden when no relevant poll) opens a bottom sheet reusing the
existing data layer and results view. New selection flow: tapping a
card only moves the single local selection (never a POST), confirm is
disabled until a choice exists, the confirmation step POSTs exactly
once (lock-ref + busy), and the server's votedOptionId locks the view
- a registered vote can never be changed. Home integration and the
old feed components are removed; helpers/results stay shared.
15 pure-helper checks green.
